* all: Eliminating warnings (mainly from Win32 builds).

+++ usecode/compiler/ucsym.cc
@@ -539,12 +539,12 @@ int Uc_function_symbol::gen_call(
     Uc_class *scope_vtbl    // For method calls using a different scope.
 ) {
 	char buf[200];
-	size_t parmcnt = aparms->gen_values(out);   // Want to push parm. values.
+	unsigned long parmcnt = aparms->gen_values(out);   // Want to push parm. values.
 	parmcnt += (method_num >= 0);       // Count 'this'.
 	if (parmcnt != parms.size()) {
 		sprintf(buf,
 		        "# parms. passed (%lu) doesn't match '%s' count (%lu)",
-		        parmcnt, get_name(), parms.size());
+		        parmcnt, get_name(), static_cast<unsigned long>(parms.size()));
 		Uc_location::yyerror(buf);
 	}
 	// See if expecting a return value from a function that has none.
